refactored save moose for save button and it stores to local storage and clears moose array
