+++ app/src/state/actions/index.ts
@@ -5,4 +5,5 @@ export const ACTIVITY_UPDATE_MOOSE = "ACTIVITY_UPDATE_MOOSE";
 export const GET_GEOLOCATION = 'GET_GEOLOCATION';
 export const USER_CLICK_RECORD_GENDER = 'USER_CLICK_RECORD_GENDER';
 export const ACTIVITY_DELETE_MOOSE = "ACTIVITY_DELETE_MOOSE";
-export const ACTIVITY_CLEAR_MOOSE_ARRAY = "ACTIVITY_CLEAR_MOOST_ARRAY";
+export const ACTIVITY_CLEAR_MOOSE_ARRAY = "ACTIVITY_CLEAR_MOOSE_ARRAY";
+export const USER_SAVE_SIGHTINGS = "USER_SAVE_SIGHTINGS";
